fixed cart

+++ src/Services/WooService.php
@@ -4,6 +4,7 @@ namespace Donapp\Services;
 
 use DateTime;
 use Donapp\Models\UserCart;
+use Exception;
 use Kernel\DB;
 
 class WooService
@@ -12,7 +13,9 @@ class WooService
     public function addToCart($data){
         
         $productId = $this->createOrUpdateProduct($data['product']);
-        
+        if(!$productId){
+            throw new Exception('Service Not Available');
+        }
         $this->deleteExpiredCarts();
         $userCart = new UserCart();
         $currentCart = $userCart->where('identifier', '=',$data['id'])->first();
@@ -67,9 +70,6 @@ class WooService
             // Update the existing product with the new data
             $product_id = $existing_product[0]['ID'];
             var_dump($product_id);
-            die();
-            $product = wc_get_product($product_id);
-    
             if ($product) {
                 $product->set_name($data['name']);
                 $product->set_regular_price($data['price']);
@@ -79,8 +79,6 @@ class WooService
                 return $product_id; // Return updated product ID
             }
         } else {
-            var_dump('new');
-            die();
             // Create a new product
             $product = new \WC_Product_Simple();
             $product->set_name($data['name']);
